+++ src/tests/frameworkVitestSetup.ts
@@ -1,0 +1,10 @@
+import { beforeAll, afterAll, beforeEach, vi } from 'vitest';
+import { createDefaultTestUser } from './testHelpers.ts';
+
+beforeAll(async () => {
+  await createDefaultTestUser();
+});
+
+afterAll(async () => {
+  // do seomthing
+});