+++ vitest.config.ts
@@ -4,7 +4,10 @@ import { defineConfig } from 'vitest/config';
 export default defineConfig({
   test: {
     globalSetup: './src/tests/globalSetupVitest.ts',
-    setupFiles: './src/tests/setupVitest.ts',
+    setupFiles: [
+      './src/tests/setupVitest.ts',
+      './src/tests/frameworkVitestSetup.ts',
+    ],
     testTimeout: 10000,
     outputFile: './coverage/junit.rspec.xml',
     reporters: ['default', 'junit'],